Generate sitemap for production
