+++ src/components/blog/posts/index.js
@@ -1,26 +1,7 @@
-const PostList = [
-  {
-    title: 'My First Blog Post',
-    filename: 'post1.md',
-    date: '2017-04-12'
-  }
-]
+import * as posts from './posts.js'
 
-// ---------------------------
-
-function Slugify (title) {
-  return title.replace(/[^a-z0-9]/gi, '-').toLowerCase()
-}
-
-function generatePostObject () {
-  var obj = {}
-  for (var post of PostList) {
-    var slug = Slugify(post.title)
-    obj[slug] = post
-  }
-  return obj
-}
-
-var PostObject = generatePostObject()
+var PostList = posts.postList
+var PostObject = posts.postObject
+var Slugify = posts.slugify
 
 export { PostList, PostObject, Slugify }
